Attempted to fix occasional hangs
It could have been when we set the offset on play to somewhere too
close/too far-past the end of the movie. Deferred instead until after
the player has determined the length of the media so that we can
gracefully advance if the movie is too close to the end.

Also made hack attempt to recycle the player instance to avoid creation
and tear-down; this may backfire if the player is leaky (or dislikes
being asked to change movies repeatedly).